Move `describe-cache` to `inspector/graph` where it belongs.
