+++ packages/breadboard/src/inspector/graph/describe-cache.ts
